smart-filtering-poc/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-filtering-poc"
version = "0.1.0"
description = "POC de filtrado inteligente de candidatos (generación, parseo, ranking, UI)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Equipo RRHH" }]
dependencies = [
  "python-docx>=0.8.11",
  "streamlit>=1.32",
  "pandas>=2.0",
  "numpy>=1.25",
  "sentence-transformers>=2.2.2",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=24.3",
  "ruff>=0.4",
]

[project.scripts]
smart-filtering = "smart_filtering.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
